nixos/logrotate: Copy configFile to /etc, remove option
When running logrotate in a user namespaced environment, such as
an nspawn container with PrivateUsers=pick, logrotate may refuse to
start as its config file is not owned by root.

Using environment.etc we can copy the file from the store and
set the proper permissions during activation.

As such, there is much less utility in being able to configure the
configFile path now, and `include` directives supported in the
program's own config schema would suffice for including extra config
outside of the NixOS module.